refactor by adding nova_scotia_types.rs
