+++ test-e2e/src/lib.rs
@@ -1,5 +1,7 @@
 extern crate core;
 
+mod nova_scotia_types;
+
 // First, configure the Mopro FFI library
 mopro_ffi::app!();
